Adjust main navbar and main image on base.html
On base.html add space between navbar elements and right align,
change color of profile nav link to black adjust size of search
box. Changed main image on background of base.html so that text
in the jumbotron on the index page can be seen clearer adjusted
the  type of font on this main jumbron on index page added
styling to learn more button.  Adjust spacing between contact
elements in the footer of the base.html page.  Add css styling for
all of the above elements to the base.css file .