Add search criteria section to the all courses
Search by Group By and location on the all
courses page.  Also added cards to display the content of the
courses.  Index.html added link to the learn more button
at the bottom of the page to bring user to all courses page.
Base css uncommented out code.  In Base.html add home link to
the logo at the top of the page.  Also added a home link to
the nav bar.  Added link to all courses to bring the user to
the all courses page.