Add navigational buttons to all courses page
Added nav buttons next, previous, home to the bottom of
the all_courses page also added a search button
under group_by and location section on this page.
Add nav buttons to the bottom of the detailed_courses page
add to basket, add to wishlist, all courses and home.
Updated the content to be displayed about the course
Removed images in media folder renamed them locally and
uploaded them again as file name were too long.  In
base.css added styling for the nav buttons on the
bottom of the page.