Add edit_courses page
Add path for edit_courses page in courses url. Add view for
edit_courses in views.py .  Also changed in courses views.py
the group_by  in request.get to Location to check if code
will execute for this link.  All courses added pagination
code at the bottom of the page as will need it later but
its commented out for the moment.  In detailed_courses added
and edit and delete link to the container.  Added edit_courses
html page and added a url for this edit_courses to the project
urls.