Create basket app
Create back app, add url for view_basket to basket urls added
a view for view_basket to views.py for basket.  Add html page
called basket.html.  Add code to this page to render the basket
if there is items in the basket else give the option to continue
browsing. Add basket app to setting.  Add basket urls to the
main project Kac_safety urls. Add link to view basket in base.html
along with code to show amount if there is items in the basket.