Register basket model in admin.py
In basket models.py change the plural name to display
Basket instead of Baskets in admin.  Added form element
for the qty in basket.html.  Also added update and remove
buttons to the basket.html page along with the js and css
so that the pointer changes when it moves over the link