Default app config added to __init__.py for purcha
In purcahse models.py line 34 updated to include 'or 0' at
the end of the line.  Purchase views.py updated to comment
out fields from the form_data section starting in line 20
line 42 in this file changed to purchase = purchase. removed
else block at line 47. Correct spelling of session in line 56
correct spelling on line 98 of successfully.  In purchase.html
removed lastname at line 100 and also the field on line 106 and
107