In contacts app renamed fname and lname fields
Renamed theses fields to fullname as stripe was not accepting
fname and lname.
Contact models.py changed fname to fullname and added a default
value also removed the second line lname. In contact forms.py
removed all reference to fname and lname and replaced with
fullname.  In contact views.py removed all reference to fname
and lname and replace with fullname.