Add mailchimpnews app
Basket basket.html change the Continue Browsing button
to btn-secondary. In settings.py added app mailchimpnews and
added section at the bottom of file for mailchimp key
and mailchimp audience id.  Added the value of these two keys to
env.py file.  In landing index.html page changed jumbotron
to a container and updated the styling on root base.css.
For mailchimpnews added the setup for the admin panel
in admin.py,  also added custom model with 3 fields,
along with urls and views.  Added two html templates
for mailchimp error.html and success.html in the
mailchimpnews templates folder. Added the mailchimpnews
subscription to the root base.html page in the footer section.