Get total amounts to appear in purchase_success
Purchase models.py changed less than sign to greater than or equal
to at line 41.  Also changed + to - at line 45 as code was
originally adding the discount rather than deducting it on the
purchase_success page.  purchase views.py changed message.success to
messages.info so that the correct confirmation of successfull purchase
will display once the user places an order.  Also passed variables
total, discount and grand_total into the context so that they be
passed to purchase_success page.  purchase_success amended the
layout of the total figures a tthe end of the page.