CSS styling of pages
basket.html for the backet contents heading removed text-center and
aligned it to the left margin and added an underline. Alter the
position of buttons at the bottom of the page by placing both
buttons on the one line and putting the make a purchase button
as the first option and continue browsing as the second option.
Also changed bag is emptpy to basket is empty.
contact_approved.html aligned heading to left and inserted underline.
Highlighted the customer name and the kac safety team in bold.
Contact.html changed the wording of the main heading to Contact Us.
Courses forms.py fixed issue with time the course was being held as
it was defaulting to the current time.
all_courses added margin left to the main heading.
detailed_courses styled main heading and moved underline in from
left margin.
settings.py change allowed hosts.
subscribe.html remove margin left from the main heading.
purchase_success.html added home button to the bottom of the page
and changed wording on Back to coourses to continue browsing.
Added styling for all of the above to base.css.