Styling for Footer
base.html alter spacing in footer so that its less cluttered and
all text centered to make it more readable.
index.html remove extra spacing at the end of the paragraph towards
the end of the page.
base.css comment out styling for row-footer1, 2 and 3.  Add padding
to row-footer2. Also added padding underneath the company name in
the footer.
Add a background color of red to the footer container so that the
footer is all red instead of a line of  red, white and red.